fix(js): fix implicit usage of localstorage
We now check if `localstorage` and `window` exist when implicitly using it (for example in `getTheme`. Here, users might not want to rely on the code running in the browser (for RSC, astro etc.)

Explicit usage of localStorage functions like `setThemeInLocalstorage` is not guarded and will throw an error on mis-use.