feat: Add basic generator implementation
First commit for #87; This one just tries to change lightness and chroma of the input color based on pre-defined steps.

There should probably be some more guards to ensure uniformness/correctness of colors. Right now the only guardrail is the fact that scss's `color.adjust` and `color.to-gamut` will clamp out-of-gamut colors.